chore(security): move token manager to foundation
