+++ src/Foundation/Security/Token/TokenManager.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Siklid\Application\Auth\Token;
+namespace App\Foundation\Security\Token;
 
 use App\Siklid\Document\AccessToken;
 use Gesdinet\JWTRefreshTokenBundle\Generator\RefreshTokenGeneratorInterface;
