+++ src/Siklid/Application/Auth/Form/UserType.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Siklid\Application\Auth\Forms;
+namespace App\Siklid\Application\Auth\Form;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
