+++ src/Siklid/Document/User.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace App\Siklid\Document;
 
 use App\Foundation\Constraint\UniqueDocument;
-use App\Siklid\Application\Auth\Token\AccessTokenInterface;
-use App\Siklid\Application\Auth\Token\HasAccessToken;
+use App\Foundation\Security\Token\AccessTokenInterface;
+use App\Foundation\Security\Token\HasAccessToken;
 use Doctrine\ODM\MongoDB\Mapping\Annotations as MongoDB;
 use Symfony\Component\Security\Core\User\PasswordAuthenticatedUserInterface as Authenticable;
 use Symfony\Component\Security\Core\User\UserInterface;
